fix(Avaliacao): arrumando erros, testes ok

+++ backend/src/main/java/com/br/integramove/api/dto/request/AvaliacaoRequestDTO.java
@@ -1,13 +1,14 @@
 package com.br.integramove.api.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AvaliacaoRequestDTO(
         LocalDate dataAvaliacao,
-        Double peso,
-        Double altura,
-        Double imc,
-        Double percentualGordura,
-        Double circuferencia
+        BigDecimal peso,
+        BigDecimal altura,
+        BigDecimal imc,
+        BigDecimal percentualGordura,
+        BigDecimal circuferencia
 ) {
 }
