+++ backend/src/main/java/com/br/integramove/application/avaliacao/BuscarAvaliacaoOutput.java
@@ -1,14 +1,15 @@
 package com.br.integramove.application.avaliacao;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record BuscarAvaliacaoOutput(
         String id,
         LocalDate dataAvaliacao,
-        Double peso,
-        Double altura,
-        Double imc,
-        Double percentualGordura,
-        Double circuferencia
+        BigDecimal peso,
+        BigDecimal altura,
+        BigDecimal imc,
+        BigDecimal percentualGordura,
+        BigDecimal circuferencia
 ) {
 }
