+++ backend/src/main/java/com/br/integramove/application/avaliacao/CriarAvaliacaoInput.java
@@ -1,13 +1,14 @@
 package com.br.integramove.application.avaliacao;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CriarAvaliacaoInput(
         LocalDate dataAvaliacao,
-        Double peso,
-        Double altura,
-        Double imc,
-        Double percentualGordura,
-        Double circuferencia
+        BigDecimal peso,
+        BigDecimal altura,
+        BigDecimal imc,
+        BigDecimal percentualGordura,
+        BigDecimal circuferencia
 ) {
 }
